refactor sanitycheck package slightly to remove winsanity
