+++ plugin/step/sanitycheck/ensure_super_user.go
@@ -1,25 +1,13 @@
 // +build !windows
 
-package winsanitycheck
+package sanitycheck
 
 /*
 	Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 */
 
 import (
-	"errors"
 	"os/user"
-)
-
-// ErrNotSuperuser is the error raised when not running as a superuser
-var (
-	ErrNotSuperuser = errors.New("not running as superuser")
-)
-
-// Unix superuser variables to allow testing
-var (
-	UnixSuperuserUsername = "root"
-	UnixSuperuserUID      = "0"
 )
 
 // EnsureSuperuser checks that we're running as superuser
