+++ plugin/step/sanitycheck/ensure_super_user_windows.go
@@ -1,6 +1,4 @@
-// +build windows
-
-package winsanitycheck
+package sanitycheck
 
 /*
 	Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
