+++ plugin/step/sanitycheck/root.go
@@ -1,5 +1,3 @@
-// +build !windows
-
 package sanitycheck
 
 /*
@@ -8,7 +6,6 @@ package sanitycheck
 
 import (
 	"errors"
-	"os/user"
 )
 
 // ErrNotSuperuser is the error raised when not running as a superuser
@@ -21,15 +18,3 @@ var (
 	UnixSuperuserUsername = "root"
 	UnixSuperuserUID      = "0"
 )
-
-// EnsureSuperuser checks that we're running as superuser
-func EnsureSuperuser(sc *SanityCheck) (FixFn, error) {
-	u, err := user.Current()
-	if err != nil {
-		return nil, err
-	}
-	if u.Username == UnixSuperuserUsername && u.Uid == UnixSuperuserUID {
-		return nil, nil
-	}
-	return nil, ErrNotSuperuser
-}
